[core] move install logic to monkey.py, don't deprecate
